Move widgets to widgets.py
